# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Meltwater Redaction API
  version: 1.0.0
  description: API for redacting document text and restoring it with a non-cryptographic key.
servers:
  - url: http://localhost:4000
paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: API is healthy.
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    example: ok
  /redactions:
    post:
      summary: Redact a document
      description: Replaces every matching keyword or phrase with XXXX and returns a reversible key.
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RedactionRequest"
            examples:
              assignment:
                value:
                  terms: "Hello world \u201cBoston Red Sox\u201d, \u2018Pepperoni Pizza\u2019, \u2018Cheese Pizza\u2019, beer"
                  documentText: "Hello world watched the Boston Red Sox and ordered beer."
      responses:
        "201":
          description: Document redacted successfully.
          headers:
            x-request-id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RedactionResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "422":
          $ref: "#/components/responses/DomainError"
  /unredactions:
    post:
      summary: Unredact a document
      description: Restores XXXX placeholders using a key generated by the redaction endpoint.
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnredactionRequest"
      responses:
        "200":
          description: Document restored successfully.
          headers:
            x-request-id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnredactionResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "422":
          $ref: "#/components/responses/DomainError"
  /documents/redactions:
    post:
      summary: Redact and store a document
      description: Redacts document text, persists the redacted document in SQLite, and indexes applied redaction terms for exact-match search.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoredRedactionRequest"
      responses:
        "201":
          description: Stored document created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoredDocumentWithKey"
        "400":
          $ref: "#/components/responses/ValidationError"
  /documents:
    get:
      summary: Search stored documents
      parameters:
        - name: redactedTerm
          in: query
          required: false
          schema:
            type: string
          description: Exact redacted term lookup. Matching is normalized to lowercase.
      responses:
        "200":
          description: Matching stored documents.
          content:
            application/json:
              schema:
                type: object
                required: [documents]
                properties:
                  documents:
                    type: array
                    items:
                      $ref: "#/components/schemas/StoredDocument"
  /documents/{id}:
    get:
      summary: Retrieve a stored redacted document
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "200":
          description: Stored redacted document.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoredDocument"
        "404":
          $ref: "#/components/responses/DocumentNotFound"
  /documents/{id}/redactions:
    get:
      summary: Retrieve redactions applied to a stored document
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "200":
          description: Applied redactions.
          content:
            application/json:
              schema:
                type: object
                required: [redactions]
                properties:
                  redactions:
                    type: array
                    items:
                      $ref: "#/components/schemas/RedactionEntry"
        "404":
          $ref: "#/components/responses/DocumentNotFound"
  /documents/{id}/unredactions:
    post:
      summary: Unredact a stored document
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [key]
              properties:
                key:
                  type: string
      responses:
        "200":
          description: Stored document restored with the provided key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnredactionResponse"
        "404":
          $ref: "#/components/responses/DocumentNotFound"
        "422":
          $ref: "#/components/responses/DomainError"
  /documents/{id}/audit-events:
    get:
      summary: Retrieve local prototype audit events for a stored document
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "200":
          description: Audit events captured by the SQLite prototype.
          content:
            application/json:
              schema:
                type: object
                required: [auditEvents]
                properties:
                  auditEvents:
                    type: array
                    items:
                      $ref: "#/components/schemas/AuditEvent"
components:
  parameters:
    RequestId:
      name: x-request-id
      in: header
      required: false
      schema:
        type: string
      description: Optional caller-provided correlation id.
    DocumentId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Stored document id.
  headers:
    RequestId:
      description: Request correlation id.
      schema:
        type: string
  responses:
    ValidationError:
      description: Request payload is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    DomainError:
      description: Redaction domain error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    DocumentNotFound:
      description: Stored document was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    RedactionRequest:
      type: object
      required: [terms, documentText]
      properties:
        terms:
          type: string
          minLength: 1
        documentText:
          type: string
    RedactionResponse:
      type: object
      required: [redactedText, key, redactions]
      properties:
        redactedText:
          type: string
          example: "XXXX XXXX watched the XXXX and ordered XXXX."
        key:
          type: string
        redactions:
          type: array
          items:
            $ref: "#/components/schemas/RedactionEntry"
    RedactionEntry:
      type: object
      required: [index, start, end, term, original]
      properties:
        index:
          type: integer
        start:
          type: integer
        end:
          type: integer
        term:
          type: string
        original:
          type: string
    UnredactionRequest:
      type: object
      required: [key, documentText]
      properties:
        key:
          type: string
          minLength: 1
        documentText:
          type: string
    UnredactionResponse:
      type: object
      required: [unredactedText]
      properties:
        unredactedText:
          type: string
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            requestId:
              type: string
    StoredRedactionRequest:
      allOf:
        - $ref: "#/components/schemas/RedactionRequest"
        - type: object
          properties:
            metadata:
              $ref: "#/components/schemas/DocumentMetadata"
    DocumentMetadata:
      type: object
      properties:
        title:
          type: string
        classification:
          type: string
        ownerId:
          type: string
    StoredDocument:
      type: object
      required: [id, title, classification, ownerId, redactedText, createdAt, updatedAt]
      properties:
        id:
          type: string
        title:
          type: [string, "null"]
        classification:
          type: [string, "null"]
        ownerId:
          type: [string, "null"]
        redactedText:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    StoredDocumentWithKey:
      allOf:
        - $ref: "#/components/schemas/StoredDocument"
        - type: object
          required: [key, redactions]
          properties:
            key:
              type: string
            redactions:
              type: array
              items:
                $ref: "#/components/schemas/RedactionEntry"
    AuditEvent:
      type: object
      required: [id, documentId, action, metadata, createdAt]
      properties:
        id:
          type: string
        documentId:
          type: [string, "null"]
        action:
          type: string
        metadata:
          type: object
        createdAt:
          type: string
          format: date-time